[MSBuild] Allow package reference to not specify target framework
Previously a PackageReference in a .nuproj project had to specify
a target framework. So to create a metadata package each package
dependency represented by the PackageReference item in the project
had to specify a target framework of "any". Now if the target
framework is not specified then "any" is assumed. When "any" is used
as a target framework the dependency is added to the .nuspec without
any associated target framework:

    <dependencies>
      <dependency id="Newtonsoft.Json" version="9.0.1" />
    </dependencies>