[MSBuild] Use correct assembly version in reference assemblies
When building the reference assembly an AssemblyInfo.cs file is
created for each reference assembly containing the version read
from the intersection assemblies.

The assembly name is also directly read from the intersection
assembly.

If any of the assembly names or versions from the intersection
assemblies do not match then a warning is reported by the MSBuild
task. A bait and switch NuGet package needs all the assemblies
to have the same assembly name and version.