[MSBuild] Fix build failing when cleaning reference assemblies
The clean target would fail if the obj/Debug intermediate directory
for the NuGet packaging project did not exist. The MSBuild task was
trying to find directories inside a directory which did not exist
which caused an exception to be thrown. Now a check is made to
ensure the directory exists first.